socket/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
bidict==0.22.1
frozenlist==1.4.1
h11==0.14.0
idna==3.6
multidict==6.0.4
python-engineio==3.13.2
python-socketio==4.6.0
simple-websocket==1.0.0
six==1.16.0
wsproto==1.2.0
yarl==1.9.4
